+++ Sample.Domain/Identity/IUserClaimRepository.cs
@@ -1,9 +1,10 @@
+using Microsoft.AspNetCore.Identity;
 using Sample.Domain.Users;
 
 namespace Sample.Domain.Identity;
 
 public interface IUserClaimRepository
 {
-    Task UpdateProfilePictureVersionClaimAsync(SampleUser user, byte version);
-    Task UpdateSlimUiClaimAsync(SampleUser user, bool slimUi);
+    Task<IdentityResult> UpdateProfilePictureVersionClaimAsync(SampleUser user, byte version);
+    Task<IdentityResult> UpdateSlimUiClaimAsync(SampleUser user, bool slimUi);
 }